pyproject.toml
[project]
name = "DyQuery"
version = "0.1.0"
description = "A nonebot2 plugin for getting the latest play record for Dynamite Explode, working well on Onebot V11 and Discord."
authors = [
    {name = "AXIS5", email = "axis5_secret@163.com"},
]
dependencies = [
    "nonebot2>=2.3.1",
    "httpx>=0.28.1",
    "nonebot-plugin-localstore>=0.7.4",
    "nonebot-plugin-orm>=0.8.3",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-adapter-discord>=1.1.6",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ["src"]

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.4.2",
    "isort>=5.13.2",
    "ruff>=0.4.6",
]

[tool.pdm.scripts]
format = { composite = ["isort . ","black . ","ruff check ."] }
test = "python -c 'print(\">>> Just \\\"pytest -W ignore -s\\\" when you complete your testsuite\")'"


[tool.black]
line-length = 90
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 90
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 90
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["C901", "T201", "E731", "E402"]

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
typeCheckingMode = "basic"

[tool.nonebot]

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]

"@local" = []
nonebot-adapter-discord = [{name = "Discord", module_name = "nonebot.adapters.discord"}]

plugin_dirs = ["src"]
builtin_plugins = []
